[admin] refactor(117) Public API 구조로 변경 및 경로 수정
