+++ apps/admin/src/entity/auth/index.ts
@@ -1,4 +1,3 @@
 export { adminLogin } from './auth.api'
 export { authKeys } from './auth.key'
-export { useAdminLoginMutation } from './auth.query'
 export { useAuthStore } from './auth.store'
