+++ apps/admin/src/shared/constant/index.ts
@@ -1,0 +1,1 @@
+export { ROUTES } from './routes'
